+++ CHANGELOG.md
@@ -56,6 +56,7 @@
 - Bug #237: Handle not found base item in access tree (@arogachev)
 - Enh #245: Handle same names during renaming item in `AssignmentsStorage` (@arogachev)
 - Chg #208: Rename `getAccessTree()` to `getHierarchy()` in `ItemsStorageInterface` (@arogachev)
+- Enh #251: Allow checking for user's roles in `ManagerInterface::userHasPermission()` (@arogachev)
 
 ## 1.0.2 April 20, 2023